add parseImageData helper method to server
track_controller/postTracks checks for embedded image data after
parsing track's metadata. If it is found, parseImageData helper
method saves image to file and returns an object containing the
path to the saved image and the images format which is then saved
under the Track models 'image' field. If no embeded image is detected,
the Track image field is saved with default data.