add: LibraryContextMenu component, LibraryRow component, jscodemod transforms
LibraryContextMenu opens on contextmenu event (right click) on
LibraryRow item and shows menu options for each track.

client/transforms directory contains transform scripts to modify
src code.

Renamed startNewQueue > playAsLastInQueue
If sameQueue is true, playAsLastInQueue adds the selected
track to the end of the queue and plays it.
If sameQueue is true, playAsLastInQueue starts a new queue and plays
the selected track.