add: Artist and Album models
postTracks now saves Artists and Album documents to DB based on
Track data. If a new Artist or Album is detected, a document is
created and saved with relevant track info. If existing artist and/or
album info is found, new info from the added track is updated to the
existing Artist/Album docs.