add: incomplete: implement refresh tokens
A refresh token, allong with JWT, is issued to the user on login and stored as an http-cookie in the client.
requireAuth checks if token has expired. If so, refreshToken checks the stored refresh token against an in-memory list (TODO: store in DB/redis?) of valid refresh tokens and looks for one whos value matches the user ID.
If a match is found, a new signed JWT is generated with the user ID stored in its payload and the refresh token is updated.